refactor: Move routes/address to routes/current
